refactor(jangar): extract whitepaper PDF proxy
