+++ docs/jangar/architecture-inventory.md
@@ -20,8 +20,8 @@ This file is generated by `bun run --cwd services/jangar docs:inventory:write`.
 | 4    | `services/jangar/src/server/atlas-store.ts`                                   | 1807 |
 | 5    | `services/jangar/src/server/torghut-simulation-control-plane.ts`              | 1724 |
 | 6    | `services/jangar/src/server/github-review-store.ts`                           | 1459 |
-| 7    | `services/jangar/src/server/torghut-whitepapers.ts`                           | 1384 |
-| 8    | `services/jangar/src/server/chat.ts`                                          | 1313 |
+| 7    | `services/jangar/src/server/chat.ts`                                          | 1313 |
+| 8    | `services/jangar/src/server/torghut-whitepapers.ts`                           | 1283 |
 | 9    | `services/jangar/src/routes/github/pulls/$owner/$repo/$number.tsx`            | 1027 |
 | 10   | `services/jangar/src/server/torghut-trading.ts`                               | 1000 |
 | 11   | `services/jangar/src/routes/torghut/trading.tsx`                              |  952 |
