+++ services/jangar/src/server/torghut-whitepapers.ts
@@ -6,6 +6,7 @@ import {
   resolveWhitepaperControlConfig,
   resolveWhitepaperStorageConfig as resolveConfiguredWhitepaperStorageConfig,
 } from './whitepaper-config'
+import { fetchWhitepaperPdfFromUrl } from './torghut-whitepaper-pdf-proxy'
 
 type NullableString = string | null
 
@@ -322,8 +323,6 @@ const MAX_LIMIT = 100
 const DEFAULT_SEARCH_LIMIT = 15
 const MAX_SEARCH_LIMIT = 50
 const WHITEPAPER_CONTROL_TIMEOUT_MS = 15_000
-const WHITEPAPER_SOURCE_MAX_REDIRECTS = 5
-const DEFAULT_WHITEPAPER_SOURCE_HOSTS = new Set(['github.com'])
 
 const asString = (value: unknown): string | null => {
   if (typeof value !== 'string') return null
@@ -375,40 +374,6 @@ const asStringArray = (value: unknown): string[] => {
 
   const single = asString(value)
   return single ? [single] : []
-}
-
-const resolveWhitepaperSourceAllowedHosts = (env: Record<string, string | undefined> = process.env) =>
-  new Set(
-    (env.JANGAR_WHITEPAPER_SOURCE_ALLOWED_HOSTS ?? '')
-      .split(',')
-      .map((entry) => entry.trim().toLowerCase())
-      .filter(Boolean),
-  )
-
-const isAllowedWhitepaperSourceHost = (hostname: string, env: Record<string, string | undefined> = process.env) => {
-  const normalized = hostname.trim().toLowerCase().replace(/\.$/, '')
-  return (
-    DEFAULT_WHITEPAPER_SOURCE_HOSTS.has(normalized) ||
-    normalized.endsWith('.githubusercontent.com') ||
-    resolveWhitepaperSourceAllowedHosts(env).has(normalized)
-  )
-}
-
-const resolveTrustedWhitepaperSourceUrl = (
-  rawUrl: string | URL,
-  baseUrl?: URL,
-  env: Record<string, string | undefined> = process.env,
-): URL | null => {
-  try {
-    const url = rawUrl instanceof URL ? new URL(rawUrl) : baseUrl ? new URL(rawUrl, baseUrl) : new URL(rawUrl)
-    if (url.protocol !== 'https:') return null
-    if (url.username || url.password) return null
-    if (url.port && url.port !== '443') return null
-    if (!isAllowedWhitepaperSourceHost(url.hostname, env)) return null
-    return url
-  } catch {
-    return null
-  }
 }
 
 const extractSourceAttachmentUrl = (uploadMetadata: unknown): string | null => {
@@ -497,72 +462,6 @@ const buildSignedPdfUrl = async (bucket: string, key: string) => {
     const client = whitepaperS3ClientCache(config)
     const command = new GetObjectCommand({ Bucket: bucket, Key: key })
     return await getSignedUrl(client, command, { expiresIn: 5 * 60 })
-  } catch {
-    return null
-  }
-}
-
-const formatPdfResponse = (
-  upstream: Response,
-  {
-    fileName,
-    source,
-  }: {
-    fileName: string | null
-    source: 'bucket' | 'source_url'
-  },
-) => {
-  if (!upstream.ok || !upstream.body) return null
-  const contentType = upstream.headers.get('content-type')?.split(';', 1)[0]?.trim().toLowerCase()
-  if (contentType !== 'application/pdf') return null
-
-  const safeName = (fileName ?? 'whitepaper.pdf').replace(/[^a-zA-Z0-9._-]+/g, '_')
-  const headers = new Headers()
-  headers.set('content-type', 'application/pdf')
-  headers.set('cache-control', 'private, max-age=60')
-  headers.set('content-disposition', `inline; filename="${safeName}"`)
-  headers.set('x-whitepaper-pdf-source', source)
-
-  const contentLength = upstream.headers.get('content-length')
-  if (contentLength) headers.set('content-length', contentLength)
-
-  return new Response(upstream.body, {
-    status: 200,
-    headers,
-  })
-}
-
-const fetchPdfFromUrl = async (
-  url: string,
-  {
-    fileName,
-    source,
-  }: {
-    fileName: string | null
-    source: 'bucket' | 'source_url'
-  },
-) => {
-  try {
-    if (source === 'source_url') {
-      let currentUrl = resolveTrustedWhitepaperSourceUrl(url)
-      if (!currentUrl) return null
-
-      for (let redirectCount = 0; redirectCount <= WHITEPAPER_SOURCE_MAX_REDIRECTS; redirectCount += 1) {
-        const upstream = await fetch(currentUrl, { redirect: 'manual' })
-        if (upstream.status >= 300 && upstream.status < 400) {
-          if (redirectCount === WHITEPAPER_SOURCE_MAX_REDIRECTS) return null
-          const location = upstream.headers.get('location')
-          if (!location) return null
-          currentUrl = resolveTrustedWhitepaperSourceUrl(location, currentUrl)
-          if (!currentUrl) return null
-          continue
-        }
-        return formatPdfResponse(upstream, { fileName, source })
-      }
-      return null
-    }
-
-    return formatPdfResponse(await fetch(url), { fileName, source })
   } catch {
     return null
   }
@@ -1365,7 +1264,7 @@ export const streamTorghutWhitepaperPdf = async (locator: TorghutWhitepaperPdfLo
   if (bucket && key) {
     const signedUrl = await buildSignedPdfUrl(bucket, key)
     if (signedUrl) {
-      const streamed = await fetchPdfFromUrl(signedUrl, {
+      const streamed = await fetchWhitepaperPdfFromUrl(signedUrl, {
         fileName: locator.fileName,
         source: 'bucket',
       })
@@ -1376,7 +1275,7 @@ export const streamTorghutWhitepaperPdf = async (locator: TorghutWhitepaperPdfLo
   const sourceUrl = asString(locator.sourceAttachmentUrl)
   if (!sourceUrl) return null
 
-  return fetchPdfFromUrl(sourceUrl, {
+  return fetchWhitepaperPdfFromUrl(sourceUrl, {
     fileName: locator.fileName,
     source: 'source_url',
   })
